fix issue where it didn't set username
when i generate a new profile in getUserRecord, i set it to their display name. but apparently some people don't have a display name??? so it was just setting it to undefined. this adds a short check for that and uses their real name if the display name doesn't exist. i hope everyone has a real name